+++ previous-versions/common/src/sdk/sdk.ts
@@ -60,9 +60,9 @@ export class SDKConfiguration {
     serverDefaults: any;
     language = "typescript";
     openapiDocVersion = "3.0.0";
-    sdkVersion = "0.40.0";
-    genVersion = "2.142.2";
-    userAgent = "speakeasy-sdk/typescript 0.40.0 2.142.2 3.0.0 @codat/common";
+    sdkVersion = "0.40.1";
+    genVersion = "2.143.2";
+    userAgent = "speakeasy-sdk/typescript 0.40.1 2.143.2 3.0.0 @codat/common";
     retryConfig?: utils.RetryConfig;
     public constructor(init?: Partial<SDKConfiguration>) {
         Object.assign(this, init);