+++ previous-versions/common/RELEASES.md
@@ -711,3 +711,13 @@ Based on:
 - [typescript v0.40.0] previous-versions/common
 ### Releases
 - [NPM v0.40.0] https://www.npmjs.com/package/@codat/common/v/0.40.0 - previous-versions/common
+
+## 2023-10-02 00:04:00
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Platform.yaml
+- Speakeasy CLI 1.92.3 (2.143.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.40.1] previous-versions/common
+### Releases
+- [NPM v0.40.1] https://www.npmjs.com/package/@codat/common/v/0.40.1 - previous-versions/common